// src/app/shared/models/config-model.ts
export class Config {
  apiUrl: string;
  apiKey: string;
  movieUrl: string;
  searchUrl: string;
  personUrl: string;
  params: string;
  imgPath: string;
  midImgPath: string;
  smallImgPath: string;
  bigBackPath: string;
  midBackPath: string;
  smallBackPath: string;
  paramsEn: string;
  imgFull: string;
  authUrl: string;
  authUrlUser: string;
  authId: string;
  userUrl: string;
  userPath: string;
  userAccId: string;
  userAccIdList: string;
  paramsApi: string;
  paramsUser:string;
  videoPath: string;
  videoPathVimeo: string;
}